restructuring for readability
